EmplyeCrud

+++ src/main/java/tn/esprit/spring/entities/Employe.java
@@ -31,7 +31,7 @@ public class Employe implements Serializable {
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
 	//@SequenceGenerator(name="seq")
-	private int id;
+	private Long id;
 	
 	private String prenom;
 	
@@ -69,7 +69,7 @@ public class Employe implements Serializable {
 	}
 	
 		
-	public Employe(int id, String prenom, String nom, String email, String password, boolean actif, Role role) {
+	public Employe(Long id, String prenom, String nom, String email, String password, boolean actif, Role role) {
 		super();
 		this.id = id;
 		this.prenom = prenom;
@@ -99,11 +99,13 @@ public class Employe implements Serializable {
 		this.role = role;
 	}
 	
-	public int getId() {
+	
+
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
@@ -190,6 +192,12 @@ public class Employe implements Serializable {
 		return "Employe [id=" + id + ", prenom=" + prenom + ", nom=" + nom + ", email=" + email + ", password="
 				+ password + ", actif=" + actif + ", role=" + role + "]";
 	}
+
+
+	
+
+
+
